feat(s2s): move to node-otp for OTP generation
