+++ package.json
@@ -8,9 +8,9 @@
     "test": "jest",
     "test:coverage": "jest --coverage",
     "build": "tsc",
-    "lint": "eslint '*/**/*.{js,ts}' --quiet --fix",
+    "lint": "eslint \"*/**/*.{js,ts}\" --quiet --fix",
     "commit": "git cz",
-    "ci:lint": "eslint '*/**/*.{js,ts}' --quiet",
+    "ci:lint": "eslint \"*/**/*.{js,ts}\" --quiet",
     "ci": "yarn && yarn ci:lint && yarn build",
     "watch": "tsc --watch"
   },
@@ -43,8 +43,6 @@
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/github": "^7.0.7",
-    "@types/chai": "^4.1.7",
-    "@types/chai-as-promised": "^7.1.2",
     "@types/express": "^4.17.2",
     "@types/express-session": "^1.15.16",
     "@types/hapi__joi": "^17.1.0",
@@ -54,11 +52,8 @@
     "@types/passport": "^1.0.2",
     "@types/passport-oauth2": "^1.4.9",
     "@types/session-file-store": "^1.2.1",
-    "@types/sinon-chai": "^3.2.4",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
-    "chai": "^4.2.0",
-    "chai-as-promised": "7.1.1",
     "commitizen": "^4.0.3",
     "commitlint-config-cz": "^0.13.1",
     "cz-conventional-changelog": "^3.0.2",
@@ -69,9 +64,6 @@
     "jest": "^24.9.0",
     "lint-staged": "^10.2.2",
     "prettier": "^2.0.5",
-    "sinon": "^7.2.3",
-    "sinon-chai": "^3.3.0",
-    "sinon-express-mock": "^2.1.0",
     "standard-version": "^7.0.1",
     "ts-jest": "^24.2.0",
     "ts-node": "^8.10.1",
@@ -91,8 +83,8 @@
     "express": "^4.17.1",
     "express-session": "^1.17.0",
     "jwt-decode": "^2.2.0",
+    "node-otp": "^1.2.2",
     "openid-client": "^3.10.0",
-    "otp": "^0.1.3",
     "passport": "^0.4.1",
     "passport-oauth2": "^1.5.0",
     "session-file-store": "^1.3.1"
