+++ src/auth/s2s/s2s.class.ts
